+++ btfxwss/client.py
@@ -43,7 +43,7 @@ class BtfxWss:
         self.secret = secret if secret else ''
 
         # Set up book-keeping variables & configurations
-        self.channel_configs = defaultdict(dict)  # Variables, as set by subscribe command
+        self.channel_configs = defaultdict(dict)
 
         self.conn = WebSocketConnection(log_level=log_level,
                                         **wss_kwargs)
@@ -56,86 +56,158 @@ class BtfxWss:
 
     @property
     def orders(self):
+        """Return queue containing open orders associated with the user account.
+        
+        :return: Queue()
+        """
         return self.queue_processor.account['Orders']
 
     @property
     def positions(self):
+        """Return queue containing open positions associated with the user 
+        account.
+        
+        :return: Queue()
+        """
         return self.queue_processor.account['Positions']
 
     @property
-    def orders(self):
-        return self.queue_processor.account['Orders']
-
-    @property
     def historical_orders(self):
+        """Return history of orders associated with the user account.
+        
+        :return: Queue()
+        """
         return self.queue_processor.account['Historical Orders']
 
     @property
-    def trades(self):
+    def transactions(self):
+        """Return history of trades associacted with the user account.
+        
+        :return: Queue()
+        """
         return self.queue_processor.account['Trades']
 
     @property
     def loans(self):
+        """Return current loans associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Loans']
 
     @property
     def historical_loans(self):
+        """Return history of loans associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Historical Loans']
 
     @property
     def wallets(self):
+        """Return wallet balances associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Wallets']
 
     @property
     def balance_info(self):
+        """Return balance information associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Balance Info']
 
     @property
     def margin_info(self):
+        """Return margin information associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Margin Info']
 
     @property
     def offers(self):
+        """Return current offers associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Offers']
 
     @property
     def historical_offers(self):
+        """Return history of offers associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Historical Offers']
 
     @property
     def funding_info(self):
+        """Return funding information associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Funding Info']
 
     @property
     def credits(self):
+        """Return current credits associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Credits']
 
     @property
     def historical_credits(self):
+        """Return history of credits associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Historical Credits']
 
     @property
     def channel_directory(self):
+        """Return channel directory of currently subscribed channels.
+
+        :return: Queue()
+        """
         return self.queue_processor.channel_directory
 
     @property
     def funding_trades(self):
+        """Return funding trades associacted with the user account.
+
+        :return: Queue()
+        """
         return self.queue_processor.account['Funding_trades']
 
     @property
     def notifications(self):
-        return self.queue_processor.account['Notifications']
+        """Return notifications associacted with the user account.
 
+        :return: Queue()
+        """
+        return self.queue_processor.account['Notifications']
 
     ##############################################
     # Client Initialization and Shutdown Methods #
     ##############################################
 
     def start(self):
+        """Start the client.
+        
+        :return: 
+        """
         self.conn.start()
         self.queue_processor.start()
 
     def stop(self):
+        """Stop the client.
+        
+        :return: 
+        """
         self.conn.disconnect()
         self.queue_processor.join()
 
@@ -144,6 +216,11 @@ class BtfxWss:
     ##########################
 
     def tickers(self, pair):
+        """Return a queue containing all received ticker data.
+        
+        :param pair: 
+        :return: Queue()
+        """
         key = ('ticker', pair)
         if key in self.queue_processor.tickers:
             return self.queue_processor.tickers[key]
@@ -151,6 +228,11 @@ class BtfxWss:
             raise KeyError(pair)
 
     def books(self, pair):
+        """Return a queue containing all received book data.
+
+        :param pair: 
+        :return: Queue()
+        """
         key = ('book', pair)
         if key in self.queue_processor.books:
             return self.queue_processor.books[key]
@@ -158,6 +240,11 @@ class BtfxWss:
             raise KeyError(pair)
 
     def raw_books(self, pair):
+        """Return a queue containing all received raw book data.
+
+        :param pair: 
+        :return: Queue()
+        """
         key = ('raw_book', pair)
         if key in self.queue_processor.raw_books:
             return self.queue_processor.raw_books[key]
@@ -165,6 +252,11 @@ class BtfxWss:
             raise KeyError(pair)
 
     def trades(self, pair):
+        """Return a queue containing all received trades data.
+
+        :param pair: 
+        :return: Queue()
+        """
         key = ('trades', pair)
         if key in self.queue_processor.trades:
             return self.queue_processor.trades[key]
@@ -172,6 +264,12 @@ class BtfxWss:
             raise KeyError(pair)
 
     def candles(self, pair, timeframe=None):
+        """Return a queue containing all received candles data.
+
+        :param pair: str, Symbol pair to request data for
+        :param timeframe: str
+        :return: Queue()
+        """
         timeframe = '1m' if not timeframe else timeframe
         key = ('candles', pair, timeframe)
         if key in self.queue_processor.candles:
@@ -192,7 +290,7 @@ class BtfxWss:
 
     def _unsubscribe(self, channel_name, identifier, **kwargs):
 
-        channel_id = self.channel_directory[(channel_name,)]
+        channel_id = self.channel_directory[identifier]
         q = {'event': 'unsubscribe', 'chanId': channel_id}
         q.update(kwargs)
         self.conn.send(**q)
@@ -220,67 +318,102 @@ class BtfxWss:
         self.conn.send(**q)
 
     @is_connected
-    def subscribe_to_ticker(self, pair, unsubscribe=False, **kwargs):
+    def subscribe_to_ticker(self, pair, **kwargs):
         """Subscribe to the passed pair's ticker channel.
 
-        :param pair: str, Pair to request data for.
+        :param pair: str, Symbol pair to request data for
         :param kwargs:
         :return:
         """
         identifier = ('ticker', pair)
-        if unsubscribe:
-            self._unsubscribe('ticker', identifier, symbol=pair, **kwargs)
-        else:
-            self._subscribe('ticker', identifier, symbol=pair, **kwargs)
+        self._subscribe('ticker', identifier, symbol=pair, **kwargs)
 
     @is_connected
-    def subscribe_to_order_book(self, pair, unsubscribe=False, **kwargs):
+    def unsubscribe_from_ticker(self, pair, **kwargs):
+        """Unsubscribe to the passed pair's ticker channel.
+
+        :param pair: str, Symbol pair to request data for
+        :param kwargs:
+        :return:
+        """
+        identifier = ('ticker', pair)
+        self._unsubscribe('ticker', identifier, symbol=pair, **kwargs)
+
+    @is_connected
+    def subscribe_to_order_book(self, pair, **kwargs):
         """Subscribe to the passed pair's order book channel.
 
-        :param pair: str, Pair to request data for.
+        :param pair: str, Symbol pair to request data for
         :param kwargs:
         :return:
         """
         identifier = ('book', pair)
-        if unsubscribe:
-            self._unsubscribe('book', identifier, symbol=pair, **kwargs)
-        else:
-            self._subscribe('book', identifier, symbol=pair, **kwargs)
+        self._subscribe('book', identifier, symbol=pair, **kwargs)
 
     @is_connected
-    def subscribe_to_raw_order_book(self, pair, prec=None, unsubscribe=False, **kwargs):
+    def unsubscribe_from_order_book(self, pair, **kwargs):
+        """Unsubscribe to the passed pair's order book channel.
+
+        :param pair: str, Symbol pair to request data for
+        :param kwargs:
+        :return:
+        """
+        identifier = ('book', pair)
+        self._unsubscribe('book', identifier, symbol=pair, **kwargs)
+
+    @is_connected
+    def subscribe_to_raw_order_book(self, pair, prec=None, **kwargs):
         """Subscribe to the passed pair's raw order book channel.
 
-        :param pair: str, Pair to request data for.
+        :param pair: str, Symbol pair to request data for
+        :param prec: 
         :param kwargs:
         :return:
         """
         identifier = ('raw_book', pair)
         prec = 'R0' if prec is None else prec
-        if unsubscribe:
-            self._unsubscribe('book', identifier, pair=pair, prec=prec, **kwargs)
-        else:
-            self._subscribe('book', identifier, pair=pair, prec=prec, **kwargs)
+        self._subscribe('book', identifier, pair=pair, prec=prec, **kwargs)
 
     @is_connected
-    def subscribe_to_trades(self, pair, unsubscribe=False, **kwargs):
+    def unsubscribe_from_raw_order_book(self, pair, prec=None, **kwargs):
+        """Unsubscribe to the passed pair's raw order book channel.
+
+        :param pair: str, Symbol pair to request data for
+        :param prec: 
+        :param kwargs:
+        :return:
+        """
+        identifier = ('raw_book', pair)
+        prec = 'R0' if prec is None else prec
+        self._unsubscribe('book', identifier, pair=pair, prec=prec, **kwargs)
+
+    @is_connected
+    def subscribe_to_trades(self, pair, **kwargs):
         """Subscribe to the passed pair's trades channel.
 
-        :param pair: str, Pair to request data for.
+        :param pair: str, Symbol pair to request data for
         :param kwargs:
         :return:
         """
         identifier = ('trades', pair)
-        if unsubscribe:
-            self._unsubscribe('trades', identifier, symbol=pair, **kwargs)
-        else:
-            self._subscribe('trades', identifier, symbol=pair, **kwargs)
+        self._subscribe('trades', identifier, symbol=pair, **kwargs)
 
     @is_connected
-    def subscribe_to_candles(self, pair, timeframe=None, unsubcribe=False, **kwargs):
+    def unsubscribe_from_trades(self, pair, **kwargs):
+        """Unsubscribe to the passed pair's trades channel.
+
+        :param pair: str, Symbol pair to request data for
+        :param kwargs:
+        :return:
+        """
+        identifier = ('trades', pair)
+        self._unsubscribe('trades', identifier, symbol=pair, **kwargs)
+
+    @is_connected
+    def subscribe_to_candles(self, pair, timeframe=None, **kwargs):
         """Subscribe to the passed pair's OHLC data channel.
 
-        :param pair: str, Pair to request data for.
+        :param pair: str, Symbol pair to request data for
         :param timeframe: str, {1m, 5m, 15m, 30m, 1h, 3h, 6h, 12h,
                                 1D, 7D, 14D, 1M}
         :param kwargs:
@@ -297,13 +430,38 @@ class BtfxWss:
         identifier = ('candles', pair, timeframe)
         pair = 't' + pair if not pair.startswith('t') else pair
         key = 'trade:' + timeframe + ':' + pair
-        if unsubcribe:
-            self._unsubscribe('candles', identifier, key=key, **kwargs)
+        self._subscribe('candles', identifier, key=key, **kwargs)
+
+    @is_connected
+    def unsubscribe_from_candles(self, pair, timeframe=None, **kwargs):
+        """Unsubscribe to the passed pair's OHLC data channel.
+
+        v
+        :param timeframe: str, {1m, 5m, 15m, 30m, 1h, 3h, 6h, 12h,
+                                1D, 7D, 14D, 1M}
+        :param kwargs:
+        :return:
+        """
+
+        valid_tfs = ['1m', '5m', '15m', '30m', '1h', '3h', '6h', '12h', '1D',
+                     '7D', '14D', '1M']
+        if timeframe:
+            if timeframe not in valid_tfs:
+                raise ValueError("timeframe must be any of %s" % valid_tfs)
         else:
-            self._subscribe('candles', identifier, key=key, **kwargs)
+            timeframe = '1m'
+        identifier = ('candles', pair, timeframe)
+        pair = 't' + pair if not pair.startswith('t') else pair
+        key = 'trade:' + timeframe + ':' + pair
+
+        self._unsubscribe('candles', identifier, key=key, **kwargs)
 
     @is_connected
     def authenticate(self):
+        """Authenticate with the Bitfinex API.
+        
+        :return: 
+        """
         if not self.key and not self.secret:
             raise ValueError("Must supply both key and secret key for API!")
         nonce = str(int(time.time() * 1000))